requirements.txt
Django==1.7.1
django-braces==1.8.1
django-flat-theme==0.9.5
Pillow==2.9.0
six==1.9.0
